I love async *SCREAMS*

+++ jobs/keyAlert.js
@@ -4,8 +4,7 @@ class KeyAlert extends RepeatedJob {
     run(bot) {
         bot.guilds.cache.each(g => {
             const settings = bot.settings[g.id];
-            if (!settings || !settings.channels.keyalerts || !settings.numerical.keyalertsage)
-                return;
+            if (!settings || !settings.channels.keyalerts || !settings.numerical.keyalertsage) {return;}
 
             const channel = bot.channels.cache.get(settings.channels.keyalerts);
             channel.messages.fetch()
@@ -13,13 +12,9 @@ class KeyAlert extends RepeatedJob {
                     bulk = [];
 
                     messages.each(message => {
-                        if (new Date() - message.createdAt > 60000 * settings.numerical.keyalertsage)
-                            bulk.push(message);
+                        if (new Date() - message.createdAt > 60000 * settings.numerical.keyalertsage) {bulk.push(message);}
                     })
-                    if (bulk.length == 1)
-                        bulk[0].delete();
-                    else if (bulk.length > 1)
-                        channel.bulkDelete(bulk);
+                    if (bulk.length == 1) {bulk[0].delete();} else if (bulk.length > 1) {channel.bulkDelete(bulk);}
                 })
         })
     }
